clients/pyproject.toml
[tool.poetry]
name = "novara"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
nats-py = "^2.13.1"
fastapi = {extras = ["standard"], version = "^0.128.4"}
jinja2 = "^3.1.6"
websockets = "^16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"